Add signup page
Made an effort to have it working right, but I don't know how
to get things working with the router, and if its just going to
redirect to login anyway. There is no way to access the signup
page at this point. There is a put request made to /signup which
might need to get messed with later.