Fix filter react issues
Filters now should be fully working. The match page is still not
very clean code, but things work at least, and filtered profiles
are showing up on the screen. For some reason everything cycles
twice which shouldn't be too hard of a fix, but maybe its just
cooler that way.